feat: add new methods and refactor logic

+++ src/app/Http/Controllers/Api/CommentController.php
@@ -22,16 +22,25 @@ class CommentController extends Controller
             'text' => $request->input('text'),
         ]);
 
+        // Загружаем связанный user
+        $comment->load('user');
+
         return response()->json($comment, 201);
     }
 
     // удалить комментарий 
     public function destroy($id)
     {
-        $comment = Comment::find($id);
+        $comment = Comment::with('image')->find($id);
 
         if (!$comment) {
             return response()->json(['message' => 'Comment not found'], 404);
+        }
+
+        $user = Auth::user();
+
+        if ($comment->user_id !== $user->id && $comment->image->author_id !== $user->id) {
+            return response()->json(['message' => 'Forbidden. You can only delete your own comments or comments to your photo.'], 403);
         }
 
         $comment->delete();
